+++ src/main/java/uk/ac/standrews/cs/population_linkage/groundTruth/AllPairsGroomMarriageBirthUmeaIdentityLinkage.java
@@ -24,7 +24,7 @@ public class AllPairsGroomMarriageBirthUmeaIdentityLinkage extends AllPairs2Sour
 
     private static final int NUMBER_OF_RUNS = 1;
 
-    public AllPairsGroomMarriageBirthUmeaIdentityLinkage(Path store_path, String repo_name, String linkage_results_filename, final String distance_results_filename, int number_of_records_to_be_checked) throws IOException {
+    public AllPairsGroomMarriageBirthUmeaIdentityLinkage(Path store_path, String repo_name, String linkage_results_filename, final String distance_results_filename, long number_of_records_to_be_checked) throws IOException {
         super(store_path,repo_name,linkage_results_filename, distance_results_filename,number_of_records_to_be_checked,NUMBER_OF_RUNS);
     }
